Fix the NAVIDROME no artist id issue (finally) and handle disc numbers
This should finally fix the Navidrome multi-artist issue. Previously we
were creating a fake id for artists that are not in the Navidrome db but
do show up on tracks because the server does not have an id for them.
But the mapping created for that item used a generic id so users saw
the last artists added this was as on every multi-artist track.

Also, now that py-opensubsonic supports disc_numbers as an attribute of
songs, give that information to MA.

Signed-off-by: Eric B Munson <eric@munsonfam.org>