Fix Don't Stop the Music for Subsonic and remove Podcast hacks
Don't stop the music was failing due to a typo when calling
get_similar_songs(), fix the typo to enable this mode.

There is a model refactor coming that will add Podcasts as first class
objects to MA so none of this code that tries to wedge them into the
Artist/Album/Track model is necessary. Tear out the dead code.

Signed-off-by: Eric B Munson <eric@munsonfam.org>