fix: Subsonic: Fix album art missing from playlist display
In order to display album art in the playlist view, we need to have an
album associated with each track. Mappings do not seem to be sufficient
here. Allow any caller of _parse_track() to optionally provide an Album
that should be used for this track and ensure that the playlist display
queries that album and uses it.

Fixes: https://github.com/music-assistant/hass-music-assistant/issues/3215

Signed-off-by: Eric B Munson <eric@munsonfam.org>