fix: Subsonic: Allow user to force player provider seek
Some Subsonic implementations advertise the ability to specify an offset
when calling the stream endpoint. This offset should allow a user to
seek by setting that value, however, at least on Gonic this value is
only honored with transcoding. Allow a user to tell the subsonic
provider that their server is broken and provider seek should be used
instead.

Fixes: https://github.com/music-assistant/hass-music-assistant/issues/3184

Signed-off-by: Eric B Munson <eric@munsonfam.org>