Fix: Correct maxsize of `MetadataLookupQueue`
Setting a default in the `_init_` hook is ineffective because the
`asyncio.Queue.__init__` always passes in a concrete value. Since that
constructor itself sets a default of zero this is what is used unless we pass
in a size at construction time.

Fixes https://github.com/music-assistant/hass-music-assistant/issues/3230.