Fix: Limit images to 512x512
My Chromecast Ultra crashes when presented with cover images on the order of
5000x5000 pixels (I've not determined the actual limit, 2048x2048 is ok
though).

This sets the default size in the player queue to 512x512 which affects all
players, as well as a chromecast specific codepath. Of the other callers of
`get_image_url` airplay already sets a size of 500x500 while the slimproto and
sonos players do not set any limit.

Fixes: https://github.com/music-assistant/hass-music-assistant/issues/3285