Fix: Deezer tracks that have been replaced with a newer version fail to play
Some tracks have been replaced by a newer version. When this happens,
the old track information stays available but it will fail to play
because the stream has become invalid. Fortunately, the playback API
provides the track replacement in a "FALLBACK" entry. So when we
encounter a track that has a "FALLBACK" entry, use the information from
that entry rather than the main track entry for the stream.

Fixes music-assistant/support#3466