fix: Subsonic: Warn when user tries to play MPEG-4 files
Because these files can store some of their metadata at the end of the
file and we can't just hand a URL to ffmpeg to pull from (due to
subsonic API) we will fail to play M4A files with the moov atom at the
back.

Add a warning in the logs when this is attempted.

Signed-off-by: Eric B Munson <eric@munsonfam.org>