Fix: Use Deezer record_type to set the album type of Deezer albums
The type property of an album in the Deezer api contains the resource
type, not the album type. The album type is called record_type instead.

Make sure to explicitly convert between types, so that we can handle
unrecognised album types as "unknown". Additionally, compilations are
called "compile" in the Deezer API so we also need to handle that.